fix teleoppolicy delay
